backend/pyproject.toml
[project]
name = "apis-as-mcp-server"
version = "0.1.0"
description = "Turn REST APIs into chat-accessible MCP tools with CopilotKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "copilotkit>=0.1.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]  # monkey-patches require imports after top-level code
"tests/*" = ["E501"]      # test files can have longer lines

[tool.ruff.lint.isort]
known-first-party = ["app", "mcp_server"]

[tool.ruff.format]
quote-style = "double"
